+++ benchmarks/ncc.py
@@ -14,7 +14,7 @@ from common import run_benchmark
 
 def main():
     cv2ext.set_log_level("DEBUG")
-    parser = argparse.ArgumentParser(description="Process iou benchmarks.")
+    parser = argparse.ArgumentParser(description="Process ncc benchmarks.")
     parser.add_argument(
         "--iterations", type=int, default=1000, help="The number of iterations to run.",
     )