+++ CHANGELOG.md
@@ -15,6 +15,11 @@
     - marlin, implements the MARLIN scheduler
       for choosing to run either a detector or cheap tracking
       algorithm to save energy
+- bboxes.match, bboxes.calculate_metrics
+    Added functions for matching and computing
+    accuracy between two lists of bounding boxes.
+- io.Display.wait and nextkey init parameters.
+    Allows more syncronization on the Display class.
 
 ### Changed